Replace TimeConverter, Add calories updating logic.

+++ app/src/main/java/com/developers/sprintsync/core/components/track/data/database/dto/TrackEntity.kt
@@ -14,7 +14,7 @@ data class TrackEntity(
     val distanceMeters: Float,
     val avgPace: Float,
     val bestPace: Float,
-    val calories: Int,
+    val calories: Float,
     val segments: Segments,
 ) {
     /*
