+++ app/src/main/java/com/developers/sprintsync/home/statistics/domain/model/TracksStatsCompact.kt
@@ -3,7 +3,7 @@ package com.developers.sprintsync.home.statistics.domain.model
 data class TracksStatsCompact(
     val runs: Int,
     val totalDistance: Float,
-    val totalCalories: Int,
+    val totalCalories: Float,
     val longestDistance: Float,
     val bestPace: Float,
     val maxDuration: Long,
@@ -13,7 +13,7 @@ data class TracksStatsCompact(
             TracksStatsCompact(
                 runs = 0,
                 totalDistance = 0f,
-                totalCalories = 0,
+                totalCalories = 0f,
                 longestDistance = 0f,
                 bestPace = 0f,
                 maxDuration = 0L,
